+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
     * version 1.2.2
     * merged most of pull request https://github.com/jvanasco/pyramid_session_redis/pull/3 from Chad Birch (@Deimos)
     * clarified assume_redis_lru in docs, added tests
+    * added `force_redis_ttl` kwarg to `RedisSession.do_refresh`
 
 -1/17/2017:
     * version 1.2.1